+++ sentinel/src/main/kotlin/com/infinum/sentinel/ui/main/preferences/PreferencesFragment.kt
@@ -8,6 +8,7 @@ import com.infinum.sentinel.data.models.raw.PreferencesData
 import com.infinum.sentinel.databinding.SentinelFragmentPreferencesBinding
 import com.infinum.sentinel.databinding.SentinelViewItemPreferenceBinding
 import com.infinum.sentinel.databinding.SentinelViewItemTextBinding
+import com.infinum.sentinel.extensions.copyToClipboard
 import com.infinum.sentinel.ui.shared.base.BaseChildFragment
 import com.infinum.sentinel.ui.shared.delegates.viewBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -17,7 +18,7 @@ internal class PreferencesFragment : BaseChildFragment(R.layout.sentinel_fragmen
 
     companion object {
         fun newInstance() = PreferencesFragment()
-        val TAG: String = PreferencesFragment::class.java.name
+        const val TAG: String = "PreferencesFragment"
     }
 
     override val binding: SentinelFragmentPreferencesBinding by viewBinding(
@@ -44,13 +45,22 @@ internal class PreferencesFragment : BaseChildFragment(R.layout.sentinel_fragmen
         SentinelViewItemPreferenceBinding.inflate(layoutInflater, binding.contentLayout, false)
             .apply {
                 nameView.text = data.name
-                data.values.forEach {
+                data.values.forEach { tuple ->
                     prefsLayout.addView(
                         SentinelViewItemTextBinding.inflate(layoutInflater, prefsLayout, false)
                             .apply {
                                 labelView.isAllCaps = false
-                                labelView.text = it.second
-                                valueView.text = it.third.toString()
+                                labelView.text = tuple.second
+                                valueView.text = tuple.third.toString()
+                                root.setOnClickListener {
+                                    println("TYPE: ${tuple.first} -> KEY: ${tuple.second} -> VALUE: ${tuple.third}")
+                                }
+                                root.setOnLongClickListener {
+                                    it.context.copyToClipboard(
+                                        key = tuple.second,
+                                        value = tuple.third.toString()
+                                    )
+                                }
                             }.root
                     )
                 }